Created LR tuner + Transformer V3
Transformer V3 is built on the ideas of ViT. My idea is that CNN's are not optimized for extracting location specific information from images. A filter in a CNN is passed over the entire image, and so specific location information can be lost. With the trajectory transformer, we want our model to be able to pay attention to features extracted from specific regions of the image since this will allow the trajectory predictor to narrow down the region it  believes the next step will occur in. To achieve this, we split our 100x100 image into 25 patches of size 20x20, and extract features from each of these patches. We then send these embeddings along with the trajectory into a modified transformer encoder, which embeds the trajectory history while also including cross attention over the image features. The goal of this is to allow this modified encoder to find an embedding for the trajectory history that best leverages the information in the image. It may also make sense to instead take this first encoder output and pass it straight to the decoder.